server/pyproject.toml
[tool.poetry]
name = "task_api"
version = "0.1.0"
description = "task_api"
authors = ["Maksim Kolyadin <whyspacy7@gmail.com>"]
package-mode = false


[tool.poetry.dependencies]
aiohttp = "^3.10.3"
aiosmtplib = "^3.0.2"
alembic = "^1.13.2"
asyncpg = "^0.29.0"
email-validator = "2.2.0"
fastapi = "^0.112.1"
fastapi-cache = "^0.1.0"
hiredis = "^3.0.0"
httplib2 = "^0.22.0"
jmespath = "^1.0.1"
mutagen = "^1.47.0"
passlib = "^1.7.4"
pydantic = "^2.8.2"
pydantic-settings = "^2.4.0"
pyjwt = "2.9.0"
python = "^3.10"
python-dotenv = "^1.0.1"
python-multipart = "^0.0.12"
requests = "^2.32.3"
sqlalchemy = "^2.0.32"
types-requests = "^2.32.0.20240712"
uvicorn = "^0.30.6"
celery = "^5.3.0"
redis = "^5.0.8"
motor = "^2.5.0"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
coverage = "^7.6.1"
flake8 = "^7.1.1"
flake8-pyproject = "^1.2.3"
httpx = "^0.27.0"
iniconfig = "^2.0.0"
mypy = "^1.11.1"
pytest = "^8.3.2"
pytest-asyncio = "^0.23.8"
pytest-dependency = "^0.6.0"
python-dateutil = "^2.9.0.post0"
python-jose="^3.3.0"
pytest-mock = "^3.14.0"
pytest-dotenv = "^0.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.flake8]
max-line-length = 200
ignore = ["E731"]
exclude =[
    ".git/",
    ".venv/",
    ".idea/",
    ".vscode/",
    "*cache*",
    "__init__.py",
]

[tool.mypy]
python_version = 3.11
warn_redundant_casts = true
warn_unused_ignores = true
disallow_subclassing_any = false
disable_error_code = [
    "unused-ignore",
    "no-any-return",
]
check_untyped_defs = false
disallow_untyped_calls = true
disallow_untyped_defs = true
warn_return_any = true
no_implicit_optional = true
strict_optional = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
